拼团抽奖 - 编辑活动 - 明细查询

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/market/groupdraw/order/OrderListParam.java
@@ -20,7 +20,7 @@ public class OrderListParam extends BasePageParam {
     private String consigneeName;
     private String mobile;
     private Timestamp createTime;
-    private Short provinceCode;
-    private Short cityCode;
-    private Short districtCode;
+    private Integer provinceCode;
+    private Integer cityCode;
+    private Integer districtCode;
 }
